requirements.txt
Flask==2.3.3
qrcode==7.4.2
Pillow==10.3.0
mysql-connector-python==8.4.0